requirements.txt
Pillow
celery
redis
django
